refactor(monitoring): move to models package
